error handling & connection
fixed with error handling e.g. for when user clicks send button without selecting receiver or tries to send a message without typing anything in the field.

Also added methods to handle registering user as online, checking if a user is online to receive message, or storing message in hashmap for now (to be saved to file).

online can now receive message, seen in terminal.