restructuring logic
thought I changed structure in messageClient in last commit but I had forgotten a lot of code in the bottom of the class. This has now been changed. I also moved logic out of the mainFrame which shouldn't happen there.

Worked on updating online list when user logs in and out. Added to server an if instanceof string, it means that user has logged out.

added a test saveUserInfo() method chain to try to understand how to save the user's contacts in the end. This has not really gotten me much.