unsentmesageswork
worked on the entity class for unsent messages as I realized it could really be good to use to send unsentMessages to client as recognizable object.

Also added some basics to the MessageHandler class which is supposed to function as a buffer to make sure the messages get sent in the correct order. Not yet implemented more than that.