fixed messages
fixed so messages are stored in a hashmap in the centerpanel waiting to be shown on the chat window. if, however, a chat window with a receiver is already open, they will see it immediately. Works well.

probably not yet implemented correctly for image messages. not tested. might need to change from jtextarea to something different to be able to show image.