can now send images
it is now possible to send and display images. logic to store these images was removed as it is not necessary.

Also improved logout logic so that we only have to send a contactsMessage and then get the owner of that message to know who logged out. These updates are only stored once the user logs out.